refactor(core): extract GovernanceActor to apps/governance
Move GovernanceActor (2,352 lines, 27 icn_governance refs) from
icn-core to a new icn-governance-actor crate in apps/governance.

Key changes:
- Add SystemEvent enum and EventEmitter trait to icn-kernel-api
- EventBus in icn-core now implements EventEmitter
- GovernanceActor takes Arc<dyn EventEmitter> instead of Arc<EventBus>
- icn-core::governance::mod.rs re-exports from icn-governance-actor
- Ratchet: 75 → 48 (-27 governance refs removed from icn-core)

This is the main payoff of the governance extraction sprint.
No behavioral changes — all existing tests pass.

Part of #859
