+++ icn/apps/governance/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "icn-governance-actor"
+version = "0.1.0"
+edition = "2021"
+description = "ICN Governance Actor — distributed decision-making for the cooperative network"
+license = "MIT OR Apache-2.0"
+publish = false
+
+[dependencies]
+# Kernel API for EventEmitter trait and SystemEvent types
+icn-kernel-api = { path = "../../crates/icn-kernel-api" }
+
+# Governance domain types (proposals, votes, domains, etc.)
+icn-governance = { path = "../../crates/icn-governance" }
+
+# Identity types (Did)
+icn-identity = { path = "../../crates/icn-identity" }
+
+# Entity types (EntityId, EntityRegistry)
+icn-entity = { path = "../../crates/icn-entity" }
+
+# Storage trait
+icn-store = { path = "../../crates/icn-store" }
+
+# Gossip actor (for governance message broadcast)
+icn-gossip = { path = "../../crates/icn-gossip" }
+
+# Observability metrics
+icn-obs = { path = "../../crates/icn-obs" }
+
+# Time utilities
+icn-time = { path = "../../crates/icn-time" }
+
+# Error handling
+anyhow = "1.0"
+
+# Async
+async-trait = "0.1"
+tokio = { version = "1", features = ["sync", "time", "rt"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+
+# Logging
+tracing = "0.1"
