+++ icn/apps/governance/src/lib.rs
@@ -1,0 +1,19 @@
+//! ICN Governance Actor
+//!
+//! This crate provides the [`GovernanceActor`] for distributed decision-making
+//! across the ICN network. It manages governance state, proposal lifecycle,
+//! voting, and outcome evaluation.
+//!
+//! ## Architecture
+//!
+//! The governance actor is a domain-specific application that sits above the
+//! kernel layer. It depends on kernel-api traits (e.g., [`EventEmitter`],
+//! [`ProtocolParameterStore`]) rather than concrete kernel implementations,
+//! maintaining clean separation between kernel mechanisms and domain semantics.
+//!
+//! [`EventEmitter`]: icn_kernel_api::events::EventEmitter
+//! [`ProtocolParameterStore`]: icn_kernel_api::protocol_params::ProtocolParameterStore
+
+pub mod actor;
+
+pub use actor::{GovernanceActor, GovernanceCommand, GovernanceConfigLite, GovernanceHandle};
