+++ icn/crates/icn-core/Cargo.toml
@@ -48,6 +48,7 @@ icn-compute = { path = "../icn-compute" }
 icn-security.workspace = true
 icn-federation.workspace = true
 icn-steward.workspace = true
+icn-governance-actor = { path = "../../apps/governance" }
 icn-coop = { path = "../icn-coop" }
 icn-community = { path = "../icn-community" }
